+++ package.json
@@ -13,6 +13,7 @@
     "@react-navigation/bottom-tabs": "^6.6.1",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/native-stack": "^6.11.0",
+    "axios": "^1.7.7",
     "react": "18.2.0",
     "react-native": "0.74.0",
     "react-native-gesture-handler": "^2.18.1",
