+++ src/components/Carousel/styles.js
