+++ src/components/Heading/styles.js
