+++ src/components/ListItem/ArrivalItem/ArrivalsItem.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+const ArrivalItem = () => {
+  return (
+    <View>
+      <Text>ArrivalItem</Text>
+    </View>
+  );
+};
+
+export default ArrivalItem;
