+++ src/components/Lists/ArrivalsList/ArrivalsList.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+const ArrivalList = () => {
+  return (
+    <View>
+      <Text>ArrivalList</Text>
+    </View>
+  );
+};
+
+export default ArrivalList;
