+++ src/components/Lists/ProductList/styles.js
@@ -2,7 +2,8 @@ import {StyleSheet} from 'react-native';
 
 const styles = StyleSheet.create({
   container: {
-    marginTop: 12,
+    flex: 1,
+    marginLeft: 10,
   },
 });
