--- src/components/ProductComponents/Product/Product.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { View, Text } from "react-native"
-
-const Product = () => {
-    return <View><Text>Product</Text></View>
-}
-
-export default Product
