+++ src/components/index.js
@@ -1,7 +1,5 @@
-export {default as Heading} from './HomeComponents/Heading/Heading';
-export {default as Welcome} from './HomeComponents/Welcome/Welcome';
-export {default as Carousel} from './HomeComponents/Carousel/Carousel';
-export {default as ProductCardView} from './ProductComponents/PoductCardView/ProductCardView';
-export {default as Product} from './ProductComponents/Product/Product';
-export {default as ProductList} from './ProductComponents/ProductList/ProductList';
-export {default as ProductDetail} from './ProductComponents/ProductDetail/ProductDetails';
+export {default as Heading} from './Heading/Heading';
+export {default as Welcome} from './Welcome/Welcome';
+export {default as Carousel} from './Carousel/Carousel';
+export {default as ProductItem} from './ListItem/ProductItem/ProductItem';
+export {default as ProductList} from './Lists/ProductList/ProductList';
