+++ src/hook/useFetch.js
@@ -1,18 +1,21 @@
 import {View, Text} from 'react-native';
 import {useEffect, useState} from 'react';
+import axios from 'axios';
 
-const useFetch = () => {
+const useFetch = ({url}) => {
   const [data, setData] = useState([]);
   const [isloading, setIsLoading] = useState(false);
   const [error, setError] = useState(null);
 
   const fetchData = async () => {
-    setLoading(true);
+    setIsLoading(true);
     try {
-      const response = await axios.get('http://localhost:3000/api/products/');
+      const response = await axios.get(url);
       setData(response.data);
       setIsLoading(false);
     } catch (error) {
+      console.log('[error]', error);
+
       setError(error);
     } finally {
       setIsLoading(false);
@@ -21,6 +24,7 @@ const useFetch = () => {
   useEffect(() => {
     fetchData();
   }, []);
+
   const refetch = () => {
     setIsLoading(true);
     fetchData();
