+++ src/screens/index.js
@@ -2,6 +2,7 @@ import Home from './Home/Home.jsx';
 import Search from './Search/Search.jsx';
 import Profile from './Profile/Profile.jsx';
 import Cart from './Cart/Cart.jsx';
-import ProductDetails from '../components/ProductComponents/ProductDetail/ProductDetails.jsx';
+import ProductDetails from './ProductDetails/ProductDetails.jsx';
+import NewRivals from './NewRivals/NewRivals.jsx';
 
-export {Home, Search, Profile, Cart, ProductDetails};
+export {Home, Search, Profile, Cart, ProductDetails, NewRivals};
